+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/rivers/InformasjonsbrevTest.kt
@@ -24,6 +24,7 @@ import no.nav.etterlatte.rapidsandrivers.BEHANDLING_ID_KEY
 import no.nav.etterlatte.rapidsandrivers.FNR_KEY
 import no.nav.etterlatte.rapidsandrivers.SAK_ID_KEY
 import no.nav.etterlatte.rivers.StartBrevgenereringRepository.Databasetabell
+import no.nav.etterlatte.rivers.opprettogjournalfoer.OpprettJournalfoerOgDistribuerRiver
 import no.nav.helse.rapids_rivers.JsonMessage
 import no.nav.helse.rapids_rivers.testsupport.TestRapid
 import org.junit.jupiter.api.Assertions.assertEquals
