+++ libs/rapidsandrivers-extras/src/main/kotlin/rapidsandrivers/Feilhaandtering.kt
@@ -1,10 +1,12 @@
 package no.nav.etterlatte.rapidsandrivers
 
+import no.nav.etterlatte.libs.common.logging.sikkerlogger
 import no.nav.etterlatte.libs.common.rapidsandrivers.feilendeSteg
 import no.nav.etterlatte.libs.common.rapidsandrivers.feilmelding
 import no.nav.etterlatte.libs.common.rapidsandrivers.setEventNameForHendelseType
 import no.nav.helse.rapids_rivers.JsonMessage
 import no.nav.helse.rapids_rivers.MessageContext
+import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 val feilhaandteringLogger = LoggerFactory.getLogger("feilhaandtering-kafka")
@@ -19,7 +21,7 @@ internal fun <T> withFeilhaandtering(
     try {
         Result.success(block())
     } catch (e: Exception) {
-        feilhaandteringLogger.warn("Håndtering av melding ${packet.id} feila på steg $feilendeSteg.", e)
+        feilhaandteringLogger.error("Håndtering av melding ${packet.id} feila på steg $feilendeSteg.", e)
         try {
             packet.setEventNameForHendelseType(EventNames.FEILA)
             packet.feilendeSteg = feilendeSteg
@@ -27,6 +29,8 @@ internal fun <T> withFeilhaandtering(
             packet.feilmelding = e.stackTraceToString()
             context.publish(packet.toJson())
             feilhaandteringLogger.info("Publiserte feila-melding")
+            val sikkerLogg: Logger = sikkerlogger()
+            sikkerLogg.error("Håndtering av melding ${packet.id} feila på steg $feilendeSteg. med body ${packet.toJson()}", e)
         } catch (e2: Exception) {
             feilhaandteringLogger.warn("Feil under feilhåndtering for ${packet.id}", e2)
         }
